refactor: image changes, padding fix, svg displaying
